♻️ [Refactoring]: #454 pdf_object.rs テストモジュールを 6 ファイルに物理分割
pdf_object.rs (989 行) 末尾の #[cfg(test)] mod tests ブロック (75 テスト・826 行) を
カテゴリ別 6 サブファイルへ物理分割し、本体を 164 行に圧縮した。

直近 #452 (lexer.rs 16 ファイル化) の 3 層構造 (本体.rs + 本体/tests.rs + 本体/tests/*.rs)
を踏襲。本リポジトリ初の「ハブ tests.rs に共通ヘルパー (make_ref / make_stream) を集約」
パターンを導入。

- pdf_object.rs L1-161 (use + enum + impl) は完全不変
- L163-989 の tests ブロックを #[cfg(test)] mod tests; の 2 行に圧縮
- tests/primitive_variants.rs (12 テスト) — Null/Boolean/Integer/Real バリアント
- tests/string_name.rs (17 テスト) — LiteralString/HexString/Name バリアント
- tests/array_dictionary.rs (21 テスト) — Array/Dictionary バリアント
- tests/reference.rs (6 テスト) — Reference バリアント
- tests/stream.rs (9 テスト) — Stream バリアント
- tests/eq_clone_debug.rs (10 テスト) — Eq/Clone/Debug 横断検証
- 75 テスト関数名と 81 件の日本語コメントを完全保持

検証:
- cargo test -p pdfmod-core: 790 passed; 0 failed
- cargo fmt -- --check: pass
- cargo clippy -p pdfmod-core --tests -- -D warnings: pass
